+++ src/main/java/mod/gottsch/forge/evercrops/core/mixin/NetherWartBlockMixin.java
@@ -17,6 +17,7 @@
  */
 package mod.gottsch.forge.evercrops.core.mixin;
 
+import mod.gottsch.forge.evercrops.core.config.Config;
 import mod.gottsch.forge.evercrops.core.persistence.CropCatchUp;
 import mod.gottsch.forge.evercrops.core.persistence.CropRegistry;
 import mod.gottsch.forge.evercrops.core.persistence.CropState;
@@ -53,6 +54,7 @@ public abstract class NetherWartBlockMixin extends BushBlock {
 
     @Inject(method = "randomTick", at = @At(value = "HEAD"))
     public void everCrops_randomTick(BlockState state, ServerLevel level, BlockPos pos, RandomSource randomSource, CallbackInfo ci) {
+        if (!Config.SERVER.bushCropsEnabled.get()) return;
         if (!state.hasProperty(NetherWartBlock.AGE)) {
             return;
         }
@@ -79,6 +81,7 @@ public abstract class NetherWartBlockMixin extends BushBlock {
 
     @Inject(method = "randomTick", at = @At(value = "INVOKE", target = "Lnet/minecraft/server/level/ServerLevel;setBlock(Lnet/minecraft/core/BlockPos;Lnet/minecraft/world/level/block/state/BlockState;I)Z"))
     public void everCrops_randomTick_setBlock(BlockState state, ServerLevel level, BlockPos pos, RandomSource randomSource, CallbackInfo ci) {
+        if (!Config.SERVER.bushCropsEnabled.get()) return;
         if (!state.hasProperty(NetherWartBlock.AGE)) {
             return;
         }