+++ src/main/java/mod/gottsch/forge/evercrops/core/mixin/StemBlockMixin.java
@@ -17,6 +17,7 @@
  */
 package mod.gottsch.forge.evercrops.core.mixin;
 
+import mod.gottsch.forge.evercrops.core.config.Config;
 import mod.gottsch.forge.evercrops.core.persistence.CropRegistry;
 import mod.gottsch.forge.evercrops.core.persistence.CropState;
 import net.minecraft.core.BlockPos;
@@ -58,6 +59,7 @@ public abstract class StemBlockMixin extends BushBlock implements BonemealableBl
 
     @Inject(method = "randomTick", at = @At(value = "HEAD"))
     public void everCrops_randomTick(BlockState state, ServerLevel level, BlockPos pos, RandomSource randomSource, CallbackInfo ci) {
+        if (!Config.SERVER.stemCropsEnabled.get()) return;
         if (!state.hasProperty(StemBlock.AGE)) {
             return;
         }
@@ -130,6 +132,7 @@ public abstract class StemBlockMixin extends BushBlock implements BonemealableBl
 
     @Inject(method = "randomTick", at = @At(value = "INVOKE", target = "Lnet/minecraft/server/level/ServerLevel;setBlock(Lnet/minecraft/core/BlockPos;Lnet/minecraft/world/level/block/state/BlockState;I)Z"))
     public void everCrops_randomTick_setBlock(BlockState state, ServerLevel level, BlockPos pos, RandomSource randomSource, CallbackInfo ci) {
+        if (!Config.SERVER.stemCropsEnabled.get()) return;
         if (!state.hasProperty(StemBlock.AGE)) {
             return;
         }